Split main file into translator and visitinfo
Keep unrelated classes in independent files.
