+++ python/lsst/obs/metadata/__init__.py
@@ -19,6 +19,7 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from .metadata import *
+from .visitinfo import *
+from .translator import *
 from .translators import *
 from .version import *
